+++ CHANGELOG.md
@@ -3,5 +3,6 @@
 ## [4.0.0](https://github.com/indexdata/ui-inventory-import/tree/v4.0.0) (IN PROGRESS)
 
 * First version of `ui-inventory-import`, initially derived from [`ui-harvester-admin` v3.0.0](https://github.com/indexdata/ui-harvester-admin/tree/v3.0.0).
+* [UIINIMP-4] Align APIs for Settings > Inventory import > Transformation steps page.
 
 (See [the `ui-harvester-admin` change-log](https://github.com/indexdata/ui-harvester-admin/blob/v3.0.0/CHANGELOG.md) for changes prior to v3.0.0.)